requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
